Add files via upload
Hibernate mappings are one of the key features of Hibernate. They establish the relationship between two database tables as attributes in your model. That allows you to easily navigate the associations in your model and Criteria queries.

You can establish either unidirectional or bidirectional i.e you can either model them as an attribute on only one of the associated entities or on both. It will not impact your database mapping tables, but it defines in which direction you can use the relationship in your model and Criteria queries.

The relationship that can be established between entities are-

One to One — It represents the one to one relationship between two tables.
One to Many/Many to One — It represents the one to many relationship between two tables.
Many to Many — It represents the many to many relationship between two tables.
You all must have heard about these relationships. In this article, you will learn about all relationships using Hibernate.